✅ Chores: 타임라인 이벤트 추가 수정
